Add tracing using the new G3Pipeline::GetCurrentModule() interface if the EventBuilder output is getting backed up.
If the number of pending output frames exceeds some large number (1000 by default), warn about potential pipeline stalls and identify the currently executing module to make it easier to trace hangs. If profiling is turned off, the code will still warn about IO stalls but will not be able to guess where they are are occurring.