Add a buffer-protocol interface to G3TimestreamMap.
This allows direct casting of a G3TimestreamMap to a 2D numpy array in a reasonably efficient way, making numpy.asarray(G3TimestreamMap) equivalent to numpy.asarray(list(G3TimestreamMap.values()), though with less looping in Python. Unfortunately, this does not extend to providing a view of the G3TimestreamMap as a result of a bug in numpy in which it passes PyBuf_INDIRECT, but does not actually support indirect buffers.