Check for too-new data structures, rather than silently crashing.
Implement a new macro (G3_CHECK_VERSION) that will automatically check
if a serialized object being loaded has a version newer than the current
version of the class. This can happen if the software is out of date and
would almost certainly recent in silent data corruption and/or segmentation
faults as the new version is generally a different length than the old one,
leading to frame shifts.

The macro can be used as follows:
template <class A>
void MyClass::serialize(A &ar, unsigned v)
{
        G3_CHECK_VERSION(v);

This will throw an exception and print an information error if v is too new.

Reviewed by:	Sasha Rahlin