Add note about strict reverse path filtering in Linux.
The Linux kernel, by default, implements an implicit bogon filter for
multicast UDP packets in which it discards all packets that appear on
an interface from which there is no route to the source address. Since
it is not possible to provide a route to the single 128x mux source
address on multiple interfaces simultaneously, this just needs to be
disabled.