Rearrange record_bolodata.py to get channel names from IceBoards and avoid importing pydfmux.
If board serial numbers are supplied as command-line arguments, then these are fed into the DfMuxBuilder.  The HousekeepingConsumer then discovers the list of boards from any Timepoint frame it receives, in order to construct both Wiring and Housekeeping frames for each one.

If no board serial numbers are supplied, then pydfmux is imported, and the list of IceBoards is determined by querying the hardware map.

Wiring frames are built from housekeeping data each time a housekeeping request is handled.

If a hardware map is not loaded, then the BolometerProperties frame is loaded from disk,
assuming it is stored in the file nominal_online_cal.g3 in the same directory as the hwm.yaml

Additionally, a --no-calibrator option is wired through to disable the CalibratorDAQ.