Make Python quaternion/angle conversion routines handle numpy arrays.
This is implemented with an in-Python parallel implementation of the
(short) C routines to exploit the numpy broadcast rules from Python.
A set of unit tests is provided to make sure that they continue doing
identical things (the old bindings to the C functions are now provided
under an alternate name).

Quaternion/angle conversions are also now in double precision.