Fix interpretation of empty frames returned by Python modules.
This addresses a truly horrible bug found by Sam Guns in which returning
an empty frame from Python would result in it being dropped. This was
particularly problematic in the case of EndProcessing frames, which
are usually empty, and will result in pipeline failures if dropped.
The bug arose because the C++ G3ModuleWrap code first checked for
modules returning lists of frames, then single frames. Since G3Frame is
iterable from Python, and an empty frame contains no elements that
cannot be cast to a G3FramePtr, it was perfectly possible to
boost::python::extract<>() a vector of frames from a single empty frame.
Since that vector (like the frame) would be empty, later code would
interpret this as a drop request, dropping the frame.

Fix this by checking for G3Frame first, and a list of frames later.