Automatically store configuration and VCS information in all output data. (#10)
Modify G3Pipeline to unconditionally include a "PipelineInfo" frame in the output that records vital statistics about when and how the processing was done (VCS branch, URL, presence of local diffs, version, running user, running host name, time of run start) and the configuration of all modules in the pipeline. To within the existence of lambda functions and the like, repr() on a PipelineInfo object can rehydrate the pipeline configuration and provide a runnable version of the script that made the file.