Add MPIAccumulator, which resembles frames back into observations.
This reassembles the frame stream from processing on N processes into
an in-memory representation resembling a TOAST TOD object, in which
frames are organized by observation, with shared metadata and a common
scan list between all processes, but node-local data that can be fetched
later.

The next piece of this is a function that fills a TOAST TOD object from
the result of MPIAccumulator.