Merge MPI support (#12)
Adds support for MPI- and TOAST-related extensions. This is made of four pieces:
* MPI-based file IO, which allows for files to be read (and processed) on multiple nodes
   in parallel. By itself, not meaningfully different from running N shared-nothing jobs, one
   on each file.
* A frame distributor, which allows an M:N IO:CPU process design, in which the data being
   read in parallel from M IO processes is analyzed by N CPU processes, with frames
   sent using MPI messages. This frame-level parallelism allows the number of nodes
   handling data to arbitrarily exceed the number of files. Metadata frames are treated
   as synchronization points.
* A tool (MPIAccumulator) to build a global view of one or more observations at the end of
   a [set of] pipeline[s] with frame-based parallelism, allowing the pipeilne to end with some
   number of arbitrary collective operations.
* An experimental tool (TOASTFiller) that uses MPIAccumulator to hand off data to TOAST
   after the completion of pipelines performing serial pre-processing.

The general model here is of a set of serial pipelines that meet at the beginning (or just after,
in the frame distributor step) and the end (in MPIAccumulator) with limited interaction in the
middle. This is designed to allow a largely serial calibration and pre-processing pipeline to be
shared between binned map-making running on shared-nothing systems and maximum-likelihood
map-making running on tightly-coupled HPC systems.

Thanks to Sasha Rahlin and Ted Kisner for review and suggestions.