G3Reader: support for setting socket timeout.
This isn't as useful as hoped, because the connection needs to be
closed if the timeout expires.  So you can't, for example, set
timeout=0 and include Process() calls in an event loop.  However, the
present solution is still usable in cases where you do not want to
hang on Process() indefinitely and when you are ok with needing to
reconnect the Reader if the timeout expires.

Full non-blocking behavior would be great to have but relies on
getting boost istream_filter to propagate WOULD_BLOCK as distinct from
EOF.  Although I haven't understood the boost stuff fully, one thing
to note is that io::file_descriptor_source definitely does not
check+propagate EAGAIN after calling read(fd).