Fix broken compiler warning about division by zero.
Division by integer zero is undefined in C, but division by floating-
point zero (which is what was meant here) is not. Depending on the
relative ordering of GCC promoting the divisor to a double from 0
and it checking for undefined behavior, /= 0 could trigger a
warning. Write it as 0.0 instead to avoid this.