Fix logic error when initializing sparse maps.
When initializing a sparse map by asking for a pixel somewhere in the map,
the map should contain one column with one element at the location of the
pixel. The code, however, saw that the current column/row offsets were (0,0),
and so filled right and down to the location of the first pixel, making
the resulting map much denser than intended. Special-case the zero-length
case to fix this.