Sparse storage for sky maps (#13)
Retool the flat-sky and Healpix map storage classes to use (optionally, but by default) a sparse backend for data storage. For non-rectangular flat-sky maps, this can reduce memory consumption by large factors. Substantial parts of the map API are also rewritten to be more useful and consistent. The Healpix map object has also been rewritten from scratch to be less baroque and to support better map storage.

The sparse backends are designed to be transparent to code: the main interface to the maps (operator []/python __getitem__()) will allocate pixels in the sparse map on assignment. Operations that require format conversions (numpy.asarray()) will implicitly adjust the map storage as needed (to dense storage in this case). C++ and Python functions are also provided to explicitly change the backend. The sparse structures, when used, exist both in RAM and on disk. Maps start sparse by default so that the map-maker can build them gradually without needing the whole sky stored in advance.

For flat-sky projections, two backends are provided: dense (as before) and dense-region sparse. The second stores a contiguous set of columns, each of which has a contiguous set of row entries. This matches the typical structure of a partial-sky map---there is a dense region in the middle surrounded by zeros---without the overhead of storing indices for every pixel. Old flat-sky maps continue to be readable. FlatSkyMap objects can now be constructed with the projection centered on an arbitrary reference pixel, ensuring that all projections are handled consistently with the WCS prescription. Each map object is also assigned a wcs attribute which can be used for plotting on WCSAxes.

Healpix has three backends: dense, indexed-sparse, and ring-sparse. Dense stores every pixel on the sky. Indexed-sparse stores a list of pixel indices and pixels (as a hash map, so log(N) access times). Ring-sparse uses a 2D structure like the flat-sky class in which each ring is a "column" and each pixel on the ring is a "row". The Healpix class has been rewritten from scratch (HealpixHitPix etc. are gone) and is not backwards compatible (no cut-sky Healpix maps stored in .g3 files are believed to exist in the wild).

The G3SkyMap base class has also been moved from core to coordinateutils, while quat and G3VectorQuat have been moved into core.

Authors: Sasha Rahlin, Nathan Whitehorn
Reviewed by: Daniel Dutcher (with formal approval), Matthew Hasselfield (with informal approval)