Optimize the Healpix ringsparse representation in memory
The default ringsparse representation divides the sphere into rings such that
index 0 on a ring corresponds to ra = 0.  This is suboptimal for maps that are
centered at ra = 0, like those collected at the South Pole.

This commit adds a boolean optional argument, shift_ra, that rearranges the
ringsparse pixelization such that ring index 0 corresponds to ra = -180 degrees.
This argument defaults to False.  Since the default representation for Healpix
maps is the indexedsparse representation, it is up to the user to set this
option appropriately.

This option can be modified by setting the shift_ra attribute, only when the
HealpixSkyMap is not in the ringsparse representation.  For a map with typical
South Pole coverage, this improves the memory footprint of the ringsparse
representation by more than a factor of 2.