Optimize the Healpix ringsparse representation in memory (#16)
The default `ringsparse` representation divides the sphere into rings such that index 0 on a ring corresponds to ra = 0. This is suboptimal for maps that are centered at ra = 0, like those collected at the South Pole.

This commit adds a boolean optional argument, `shift_ra`, that rearranges the `ringsparse` pixelization such that ring index 0 corresponds to ra = -180 degrees. This argument defaults to `False`. It is up to the user to set this option appropriately, either as an argument to the constructor or by setting the `shift_ra` attribute of an existing map object. When reading in explicitly indexed HEALpix data (either from a fits file or creating a `HealpixSkyMap` instance from `numpy` arrays by hand), this argument is determined empirically from the distribution of input pixels in phi to optimize memory allocation.

For a map with typical South Pole coverage, this improves the memory footprint of the `ringsparse` representation by more than a factor of 2.

The default map representation for new maps or maps loaded from explicitly indexed fits files has also been changed from the `indexedsparse` representation to the `ringsparse` representation. Finally, setting any of the representation properties to `False` now raises an error, since that is ill-defined behavior.