G3Reader: support for setting socket timeout. (#14)
* G3Reader: support for setting socket timeout.

This isn't as useful as hoped, because the connection needs to be
closed if the timeout expires.  So you can't, for example, set
timeout=0 and include Process() calls in an event loop.  However, the
present solution is still usable in cases where you do not want to
hang on Process() indefinitely and when you are ok with needing to
reconnect the Reader if the timeout expires.

Full non-blocking behavior would be great to have but relies on
getting boost istream_filter to propagate WOULD_BLOCK as distinct from
EOF.  Although I haven't understood the boost stuff fully, one thing
to note is that io::file_descriptor_source definitely does not
check+propagate EAGAIN after calling read(fd).

* G3Reader: release the python GIL during I/O.

This improves Python threading performance, especially for Network
readers.

* G3NetworkSender: add Close method.

This is a means to shut down a Sender gracefully, since it's not
always possible to trigger the destructor reliably (ahem python).

* Use .none instead of .None in python G3FrameType

Previously, G3FrameType.None was inaccessible in Python because of the
reservedness of None.  Now you can use G3FrameType.none.  It's likely
that the only code this will break is if folks were testing, like,
str(my_frame.type) == 'None'.