Improved project versioning (#22)
 * Use `git describe` to store a sequential version string in `spt3g/version.py`
 * Make `spt3g.__version__` a string variable containing the output of `git describe`
 * Add package version info to `cmake/Spt3gConfigVersion.cmake` for use by derived cmake projects
 * Add `SPT3G_VERSION` compiler definition for access to the version in C++ code

For this versioning to work properly, the project should be tagged with a git tag of the form "v<major>.<minor>.<patch>".
If the project is not tagged, the `__version__` string falls back to using a git hash, and the cmake `Spt3gConfigVersion.cmake` file is not created.

To check the version in a cmake project, use:

    find_package(Spt3g 0.1.0)

This will check that the Spt3g package has version 0.1.0 or greater.

Versioning also works correctly when building from an exported source tree.   To configure the git repository to automatically populate the version on export, run the following command in the source directory:

    cmake/config_export.sh

And then export the source code with `git archive`.  All of the scripts that populate the version files in the build directory know how to correctly parse the VERSION file in the source tree.

This PR fixes #20