Allow streaming arcfile data from a network source (e.g. GCP) (#29)
This change allows the ARCFileReader to inject GcpSlow frames directly into a
G3Pipeline by connecting to a running GCP session.  Use "tcp://host:port" to
connect.  The port should be the CP_MONITOR_PORT as defined in
gcp/bin/control/ports.h (e.g. 5482 for SPT).

The network protocol differs slightly from the file protocol.  A handshake is
required over the TCP connection to initiate the register stream.  Further
improvements to the code could allow adjusting the contents of the received
frames, and the interval (in source frames) between them.  By default, all
available registers are sent at the source frame rate.  This includes unarchived
registers that appear in the live stream but are never written to arcfiles.