Fix build with Clang 9.
Out-of-class C++ operators need to be defined as part of the same namespace
as the class on which they operate. Since quat is a typedef of a class in
boost::math, the operator for ~ needs to be in boost::math too.

This fixes compilation, but leaves the issue that this use of operator ~,
while idiomatic in Python, is non-idiomatic in C++. We should probably
rename the C++ version of this function for G3VectorQuat (it is already
defined as conj() for quat by Boost) and retain just Python ~. For now,
just get things building.