In-place Sort() method for G3TimesampleMap objects (#33)
This method is useful when the contents of a G3TimestreamMap object are
unordered in time, e.g. when constructed from out-of-order packets.

All sorting operations are done on one vector at a time with replacement, to
minimize the required memory footprint.

Also includes some API fixes like C++ shared pointers and lower case python methods.