Fix implicit instantiation of std::vector<intXX_t> from numpy arrays.
This is a long and winding tale, starting with the fact that elements of
integer numpy arrays are not actually Python integers, but changeling types
that bear a vague, but misleading, resemblance to Python integers. Because
these elements could not be converted into C++ types by
boost::python::extract(), they were failing the convertability check in
the from-Python converter being registered for STL containers. This
prevented their implicit conversion at all. Solve this by registering
a custom from-Python converter that checks for objects that implement the
Python buffer protocol as well as the existing check for iterability with
convertible elements.

I am now going to have a beer.