Fix indexing of longs.
I had assumed/hoped that the authors of the Python struct package would
have avoided one of the core founding mistakes of C: having all the base
types be defined to be of variable, machine-dependent length. In that world,
the default type suite flag would have been "=" and not "@". In the world
we actually live in, however, the default is "@" and a long means anything
you want it to mean. Roll with the punches.