New map maker (#40)
This branch adds a new map binner with a more streamlined interface. The code is broken into four map-binning modules:

*  SingleDetectorBoresightBinner: creates (unpolarized) maps for each bolometer in boresight coordinates
*  SingleDetectorMapBinner: creates (unpolarized) maps for each bolometer in regular (RA/dec etc.) coordinates
*  MapBinner: creates one map from all the detectors (polarized or unpolarized), optionally using OpenMP to parallelize the binning by detector.
*  MapMockObserver: creates TOD for all detectors in the BolometerPropertiesMap given an input sky and scan frames containing pointing.

The main goal is to streamline the map-making flow, simplifying configuration and improving performance. Very little that changes any existing code or scripts is actually changed by this patch set -- mostly it just adds modules.

The only real change besides that is the addition of a new core class, G3TimestreamQuat, that is identical to (and inherits from) a G3VectorQuat except that it has some timestream-like properties (start, stop, sample_rate, etc.). This simplifies the use of MapMockObserver and is more generally useful anyway. The coordinate-system modules that make quaternion pointing in maps/ have been updated to emit G3TimestreamQuat instead of G3VectorQuat, and a module (maps.AddTimingToPointingQuats) has been added that updates old data using timing from other objects in the frame. Because G3TimestreamQuat inherits from G3VectorQuat, updates to the data will not break any existing code and all modules except MapMockObserver continue to expect data in the lowest-common-denominator G3VectorQuat and so work with data in either format. Parts of the C++ API that translate between G3Timestream for scaling pointing and quaternions should be updated in future, but have not yet been.

Co-authored-by: Sasha Rahlin <arahlin@fnal.gov>