Unit-agnostic Mueller matrix inverse computation
Uses a threshold on the condition number to determine whether the Mueller matrix
inverse is singular, rather than a threshold on the determinant.  The condition
number is a unitless measure, and is therefore more robust than the determinant
for this purpose.

Also avoid a unit-dependent check on off-diagonal matrix elements in computing
the condition number.

This change should only affect pixels with poor angular coverage, which are
likely to be masked anyway.