Set up testing infrastructure (#43)
Use GitHub actions to automatically try a build and run the test suite on any push (master or to a branch). This should provide better post-commit alerting of breakage and improve code review on pull requests (since the tests will automatically be run when the pull request is made).

To test portability, runs on a few of GitHub's very limited selection of platforms: Ubuntu 18 and 20. I would like to get MacOS in there, too, but there is some bad interaction between Homebrew's cmake and Boost that I can't figure out right now. Once that's resolved, we should get MacOS builds as well.