Let G3Frame::Get<>() return NULL if the requested object does not exist.
This matches the original intended behavior, and the documentation. At
some point, the behavior of Get<>() was changed to log_fatal() if the
object in question under did not exist, could not be decoded, or would
not cast to the right type as a sop to poorly-written code that did
not check the return value of Get<>() and then would crash with a much
less informative error.

Here, add a second argument to request either a log_fatal() on failure
(the default for compatibility and because it is a genuinely useful
feature) or a null/empty shared pointer, which can be used for error
checking in the calling code without multiple rounds of Has<>().