Request NULL pointers on failure of G3Frame::Get<>().
This code already handled errors internally, so we'd rather not have a
log_fatal() on failure. In particular, some of the TOD filtering code
produces an absent G3TimestreamMap as output if no detectors survive
cuts in a particular group (a wafer went bad, for instance), which should
be interpreted as a non-fatal data error (log_error) rather than a
fatal configuration error (which G3Frame was doing before).