Correct handling of patch extraction/insertion with flat sky G3SkyMapWeights objects (#53)
The FlatSkyMap methods `insert_patch`, `extract_patch` and `reshape` return
FlatSkyMap objects or take FlatSkyMap objects as arguments, so these do not
behave as expected when called as monkeypatched attributes of a G3SkyMapWeights
object.  This PR wraps these methods so that the inputs and outputs are
G3SkyMapWeights objects, with each map attribute handled appropriately by the
corresponding FlatSkyMap method.