Use the Python built-in repr() rather than calling __repr__() on objects.
For unitialized classes, __repr__() will try to call the instance method,
which will throw an exception, rather than repr-ing the class itself. This
logic is complicated and irritating to implement, so just use Python's own
repr() directly through the C API.

Additionally, do a better job of catching and alerting the user to repr()
failures.