Add support in the MapBinner for per-scan (or sub-obs) map-making.
This adds a new parameter (map_per_scan) to the map binner that can
control when maps are emitted. If defaults to False, preserving existing
behavior. If set to True, or to a callable that returns True, it will
result in the in-progress map being flushed to the data stream. This
enables simple creation of per-scan maps (set to True) or more complex
bundlings (with a callable).