Redydrate FLAC-encoded timestreams as 32-bit floats.
Both the FLAC encoder and the DfMux firmware have a maximum bit depth of
24 bits, which is the same as the bit depth of the significand of an
IEEE 754 32-bit float. As such, any number that can be encoded with FLAC
can be represented losslessly with a 32-bit float -- as can the NaNs we
encode with the data, which cannot be stored in an int32_t.

Note that care needs to be taken to prevent roundoff error when adding
these timestreams.