Integrate non-double types and array contiguity to timestreams and timestream maps (#80)
This is the first step toward a G3SuperTimestream-like interface for G3TimestreamMap and provides most of the features of the SO G3SuperTimestream class in the base G3Timestream/G3TimestreamMap. Highlights of the changes:

* A G3TimestreamMap can now be treated as a 2D array as a read/write object in numpy.
* A G3TimestreamMap can be compactified into a contiguous 2D block (necessary for above) when needed.
* A G3TimestreamMap can be initialized from a 2D numpy array and a list of keys, optionally sharing the numpy array's underlying data buffer.
* Both raw timestreams and timestream maps now support a variety of integer and floating-point data types.
* FLAC-compressed timestreams are now deserialized as floats instead of doubles, which can be done without data loss since the bit depth of the significand of an IEEE 754 32-bit floating-point number is identical to the bit depth of FLAC (24 bits). Note that care now needs to be taken to prevent roundoff error when adding these timestreams.

No breaking changes are made to the API for either G3Timestream or G3TimestreamMap.

Other changes introducing new features, in particular likely updates to the on-disk format to exploit some of the efficiencies possible by treating G3TimestreamMaps as a block, will follow in future commits.

Thanks to Sasha Rahlin and Matthew Hasselfield for code and discussions.