Area-preserving bilinear interpolation for Healpix maps (#89)
The weights computed for bilinear interpolation are computed in (theta, phi) coordinates, such that the effective sample location is shifted north or south compared to its actual location, with a shift amplitude that scales like 1/nside and is largest near the poles.  This PR changes the GetInterpPixelsWeights method to compute weights in (z, phi) coordinates, so that the weights are area-preserving and calculated in the coordinates in which the Healpix pixel boundaries are defined.

Note that this means that `HealpixSkyMap.get_interp_values()` no longer produces the same results as `healpy.get_interp_val()`.

Thank you to Eric Hivon for identifying the issue.