Fix instantiation of G3Timestreams from iterables.
Under certain circumstances that I have not quite tracked down, Boost's
type inference for the stub push_back() failed, resulting in data
corruption. Instead, use a type it is designed to support (std::vector<double>)
and retire the stub push_back() hack. This incurs a memcpy() but the
overhead is tiny given that this is only called in circumstances that
involve a Python loop anyway.

Closes #93.