Print core logging messages with timestamps by default
Timestamps on core logging messages are very useful for rudimentary
profiling, especially for long running scripts. While the profile option for
G3Pipeline.Run() provides similar functionality, it's only useful for scripts
that are entirely pipeline-based, which is often not the case. Moreover,
timestamps on log messages make it easy to track the progress of a script in
real-time, rather than waiting until it has finished running to find out how
long it took.

Closes #50.