Add frame-level bzip2 compression and decompression (#112)
This PR integrates bzip2 support to be on par with gzip support for G3 file I/O. Files with the ".g3.bz2" extension will run through the appropriate bz2 algorithm on read and write.  Bzip2 support is optional and automatically detected by cmake, but can be manually overridden using a cmake argument.