Track source filename for G3Frame objects that are read in using G3Reader (#118)
Add a ``G3Frame._filename`` attribute to frames, set when read in with G3Reader (or ARCFileReader).  This simplifies keeping track of input filenames further down the pipeline.  Documentation has been added to clarify that the attribute is fragile because it is not serialized, and therefore does not survive pickling or passing around in a multiprocessing situation.