Healpix maps with nside > 8192 (#120)
This PR ensures that maps are consistently indexed using 64-bit integers everywhere, to avoid truncation e.g. when indexing maps with high nside (>8192).  The healpix C bindings are moved out of the public API, and the header is modified to use the `ssize_t` integer type for all 64-bit functions.

Fixes #119