Make FlattenPol set flat_pol attribute of T maps and weights (#136)
`FlattenPol` sets the `flat_pol` attribute of QU maps and weights to keep track of whether a frame has been flattened, but does not set this attribute for T maps and weights. While `FlattenPol` does not affect T, a frame's T map is often used as a "stub" containing metadata for the generation of subsequent data products. As a result, polarized data products generated from a T stub will not inherit the `flat_pol` attribute as one might expect.  To avoid this issue, make `FlattenPol` modify `flat_pol` for T as well as Q and U.