Disable FITS image compression by default
Compression of floating point data using the GZIP_2 fits compression algorithm
is *not* lossless.  To avoid pitfalls with this algorithm, compression is
disabled by default.  Also expose the `quantize_level` keyword argument to
enable adjustment of the compression algorithm, if compression is desired.

See the astropy FITS image documentation for further information:
https://docs.astropy.org/en/stable/io/fits/api/images.html

Closes #135.