Allow extra calibration frames to follow proper calframe files in pipelines
Calibration frame files often have PipelineInfo frames in them for bookkeeping
purposes.  This fix allows multiple calibration frames to be placed into a
pipeline in any order, so that they can be merged properly.  This assumes that
all of the calibration frames are in order at the beginning of the pipeline,
with only PipelineInfo frames potentially interleaving them.