Smarter handling of map polarization properties (#133)
Add TT, TQ, etc to the `MapPolType` enumerator, and set the `pol_type` attribute for each map weights item.  Add a new `SetPolConv` pipeline module for changing polarization convention.  Additionally, the TU and QU attributes of a `G3SkyMapWeights` object are multiplied by -1 (along with the U map) if their polarization convention is changed from IAU to COSMO or vice versa using the `SetPolConv` module.  The `G3SkyMap.pol_conv` attribute is now a simple public attribute, and will not result in changing the sign of U/TU/QU maps when updated directly.

Where possible, determine parameters from the input map stub, similarly to the way these are handled in the `MapBinner` pipeline module.  Most notably, the `G3SkyMapWeights` constructor now determines its polarization state based on the `pol_conv` property of the input reference map.  Similarly, the `InjectMapStub` pipeline module determines whether the output frame is weighted and/or polarized based on the properties of the input stub.