G3PythonContext class for handling the Python GIL
This PR creates a new class that simplifies initialization of python threads, as
well as acquiriing / releasing the Python global interpreter lock in various
contexts.

Use cases include:

1. Ensuring that Py_Initialize() is properly called at the beginning of a
   program that is expected to interact with the python interpreter, and also
   that Py_Finalize() is called when the program is finished.
2. Ensuring that the current thread state is saved and the GIL released as
   necessary, e.g. for IO operations, and then the thread state is restored on
   completion.
3. Ensuring that the GIL is acquired for one-off interaction with the python
   interpreter, and released when complete.

A G3PythonContext object is used throughout the library code for cases 2 and 3.
If the python interpreter has not been initialized (i.e. the compiled program is
expected to be purely in C++), then these context objects are essentially no-op.
If the python interpreter is initialized (e.g. inside a python program or
command-line interface), then these context objects will handle the GIL
appropriately.

See the examples/cppexample.cxx C++ program for a simple implementation of the
above behavior.

This PR also adds logic throughout the G3PipelineInfo and G3ModuleConfig
class definitions to enable them to serialize appropriately in a pure-C++
program.