Avoid calling into the Python interpeter in the G3ModuleConfig backend
Use a G3MapFrameObject storage structure for the module arguments, rather than a
map of python objects.  Since the serialization process requires a call to
repr() for non-G3FrameObjects anyway, do this step in the python shim that
creates the config in the first place.

Also ensure that simple scalar values are serialized as frame objects.

Adds a new ``spt3g.core.to_g3frameobject`` function for converting
python objects to G3FrameObjects.