Optionally disable pipeline halt on SIGINT
Optional argument `signal_halt` for the `G3Pipeline.Run()` method.  By default,
this is set to True to enable current behavior, namely that a pipeline is halted
(via the `halt_processing` flag) and terminates after the the current frame
reaches the end of the pipeline.  If False, this signal handling is disabled,
and it is then up to the user to ensure that a running pipeline is properly
halted on interrupt.