G3Writer as a python context (#158)
This PR enables using G3Writer and G3MultiFileWriter as a python context object,
so that the object can be created and cleanly closed using the python `with`
statement.  For example:

    with core.G3Writer(filename="path/to/file.g3") as writer:
        writer(frame1)
        writer(frame2)

This will open the output file, write the two frames to it, then properly close
the file by appending an EndProcessing frame.