Build Python wheels (#159)
This PR rearranges this code to facilitate building and installing the `spt3g` python package with standard python tools like `pip` and `build`.

The shared libraries, headers, and cmake configuration files are installed relative to `CMAKE_INSTALL_PREFIX` as usual.  The `pip` installer handles versioning, and passes appropriate cmake definitions to the cmake builder that compiles all of the libraries and modules, then installs the python package into the appropriate location, and triggers the cmake installer.

The `pip` installer can also be run in an `--editable` mode that works much like the cmake build directory, enabling code development without the need to make changes to various path environment variables.  It also should install all of the components into appropriate locations when run in a virtual environment.

Versioning is done using `setuptools_scm` with the `pip` build system, or with a custom shell script with the `cmake` build system.  In both cases, the version information is determined by running a series of `git` commands or by parsing the `.git_archival.txt` file, then output to `spt3g/version.py` for access in the Python package and to the `SPT3G_VERSION` preprocessor directive for access in C++.

Finally, this PR includes a github workflow for producing pre-compiled wheels for Linux and OSX platforms.